+++ prefect_aws/__init__.py
@@ -4,6 +4,15 @@ from .client_parameters import AwsClientParameters
 from .s3 import S3Bucket
 from .ecs import ECSTask
 from .secrets_manager import AwsSecret
+from .workers import ECSWorker
+
+from prefect._internal.compatibility.deprecated import (
+    register_renamed_module,
+)
+
+register_renamed_module(
+    "prefect_aws.projects", "prefect_aws.deployments", start_date="Jun 2023"
+)
 
 __all__ = [
     "AwsCredentials",
@@ -12,6 +21,7 @@ __all__ = [
     "S3Bucket",
     "ECSTask",
     "AwsSecret",
+    "ECSWorker",
 ]
 
 __version__ = _version.get_versions()["version"]