+++ prefect_aws/s3.py
@@ -3,19 +3,18 @@ import asyncio
 import io
 import os
 import uuid
-import warnings
-from pathlib import Path, PurePath
+from pathlib import Path
 from typing import Any, BinaryIO, Dict, List, Optional, Union
-from uuid import uuid4
 
 import boto3
 from botocore.paginate import PageIterator
+from botocore.response import StreamingBody
 from prefect import get_run_logger, task
 from prefect.blocks.abstract import ObjectStorageBlock
 from prefect.filesystems import WritableDeploymentStorage, WritableFileSystem
 from prefect.utilities.asyncutils import run_sync_in_worker_thread, sync_compatible
 from prefect.utilities.filesystem import filter_files
-from pydantic import Field, root_validator, validator
+from pydantic import Field
 
 from prefect_aws import AwsCredentials, MinIOCredentials
 from prefect_aws.client_parameters import AwsClientParameters
@@ -241,14 +240,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
 
     Attributes:
         bucket_name: Name of your bucket.
-        minio_credentials: [DEPRECATED; use the credentials field instead]
-            A block containing your credentials (choose this or AWS Credentials)
-        aws_credentials: [DEPRECATED; use the credentials field instead]
-            A block containing your credentials (choose this or MinIO Credentials)
-        basepath: [DEPRECATED; use the bucket_folder field instead]
-            The base path to use when storing data in this bucket.
-        endpoint_url: [DEPRECATED; pass AwsClientParameters in AwsCredentials instead]
-            The URL of the S3-compatible object storage service.
         credentials: A block containing your credentials to AWS or MinIO.
         bucket_folder: A default path to a folder within the S3 bucket to use
             for reading and writing objects.
@@ -256,51 +247,15 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
 
     _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/1jbV4lceHOjGgunX15lUwT/db88e184d727f721575aeb054a37e277/aws.png?h=250"  # noqa
     _block_type_name = "S3 Bucket"
+    _documentation_url = (
+        "https://prefecthq.github.io/prefect-aws/s3/#prefect_aws.s3.S3Bucket"  # noqa
+    )
 
     bucket_name: str = Field(default=..., description="Name of your bucket.")
 
-    # TODO: remove deprecated minio_credentials after March 27, 2023
-    minio_credentials: Optional[MinIOCredentials] = Field(
-        default=None,
-        description=(
-            "[DEPRECATED; use the credentials field instead] "
-            "A block containing your credentials (choose this or "
-            "AWS Credentials)"
-        ),
-    )
-
-    # TODO: remove deprecated aws_credentials after March 27, 2023
-    aws_credentials: Optional[AwsCredentials] = Field(
-        default=None,
-        description=(
-            "[DEPRECATED; use the credentials field instead] "
-            "A block containing your credentials (choose this or "
-            "MinIO Credentials)."
-        ),
-    )
-
-    # TODO: remove deprecated basepath after March 27, 2023
-    basepath: Optional[Union[str, Path]] = Field(
-        default="",
-        description=(
-            "[DEPRECATED; use the bucket_folder field instead] "
-            "A default location to write to and read from in the S3 bucket. Defaults "
-            "to the root of the bucket."
-        ),
-    )
-
-    # TODO: remove deprecated endpoint_url after March 27, 2023
-    endpoint_url: Optional[str] = Field(
-        default=None,
-        description=(
-            "[DEPRECATED; pass AwsClientParameters in AwsCredentials instead] "
-            "URL endpoint to use for S3 compatible storage. Defaults to "
-            "standard AWS S3 endpoint."
-        ),
-    )
-
-    credentials: Optional[Union[AwsCredentials, MinIOCredentials]] = Field(
-        default=None, description="A block containing your credentials to AWS or MinIO."
+    credentials: Union[AwsCredentials, MinIOCredentials] = Field(
+        default_factory=AwsCredentials,
+        description="A block containing your credentials to AWS or MinIO.",
     )
 
     bucket_folder: str = Field(
@@ -311,85 +266,25 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         ),
     )
 
-    @validator("basepath", pre=True)
-    def cast_pathlib(cls, value):
-        """
-        If basepath provided, it means we aren't writing to the root directory
-        of the bucket. We need to ensure that it is a valid path. This is called
-        when the S3Bucket block is instantiated.
-        """
+    class Config:
+        smart_union = True
 
-        if issubclass(value.__class__, PurePath):
-            return value.as_posix()
-        return value
-
-    @validator("basepath", pre=True)
-    def deprecate_basepath(cls, value):
+    # Property to maintain compatibility with storage block based deployments
+    @property
+    def basepath(self) -> str:
         """
-        Raises deprecation warning.
-        """
-        if value:
-            warnings.warn(
-                "The basepath field is deprecated and will be removed March 2023. "
-                "Please use bucket_folder instead.",
-                DeprecationWarning,
-            )
-        return value
+        The base path of the S3 bucket.
 
-    @validator("endpoint_url", pre=True)
-    def deprecate_endpoint_url(cls, value):
+        Returns:
+            str: The base path of the S3 bucket.
         """
-        Raises deprecation warning.
-        """
-        if value:
-            warnings.warn(
-                "The endpoint_url field is deprecated and will be removed March 2023. "
-                "Please pass it in AwsClientParameters and pass it into "
-                "the credentials block instead.",
-                DeprecationWarning,
-            )
-        return value
+        return self.bucket_folder
 
-    @root_validator(pre=True)
-    def check_credentials(cls, values):
-        """
-        Ensure exactly 1 of 2 optional credentials fields has been provided by
-        user.
-        """
-        creds_exist = bool(values.get("credentials"))
-        minio_creds_exist = bool(values.get("minio_credentials"))
-        aws_creds_exist = bool(values.get("aws_credentials"))
-
-        # if both credentials fields provided
-        if minio_creds_exist and aws_creds_exist:
-            raise ValueError("Only one set of credentials should be provided.")
-
-        if minio_creds_exist:
-            # raise deprecationwarning
-            warnings.warn(
-                "The minio_credentials field is deprecated and "
-                "will be removed March 2023. Please use credentials instead.",
-                DeprecationWarning,
-            )
-            values["credentials"] = values.get("minio_credentials")
-        elif aws_creds_exist:
-            # raise deprecationwarning
-            warnings.warn(
-                "The aws_credentials field is deprecated and "
-                "will be removed March 2023. Please use credentials instead.",
-                DeprecationWarning,
-            )
-            values["credentials"] = values.get("aws_credentials")
-
-        # if neither credentials fields provided
-        if not creds_exist and not minio_creds_exist and not aws_creds_exist:
-            raise ValueError(
-                "S3Bucket requires at least one credentials block to be provided."
-            )
-        return values
+    @basepath.setter
+    def basepath(self, value: str) -> None:
+        self.bucket_folder = value
 
     def _resolve_path(self, path: str) -> str:
-
         """
         A helper function used in write_path to join `self.basepath` and `path`.
 
@@ -399,67 +294,34 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
                 bucket has a unique key (or key name).
 
         """
-
-        path = path or str(uuid4())
-
-        bucket_folder = self.bucket_folder or self.basepath
-        # If basepath provided, it means we won't write to the root dir of
+        # If bucket_folder provided, it means we won't write to the root dir of
         # the bucket. So we need to add it on the front of the path.
         #
         # AWS object key naming guidelines require '/' for bucket folders.
         # Get POSIX path to prevent `pathlib` from inferring '\' on Windows OS
-        path = (Path(bucket_folder) / path).as_posix() if bucket_folder else path
+        path = (
+            (Path(self.bucket_folder) / path).as_posix() if self.bucket_folder else path
+        )
 
         return path
 
     def _get_s3_client(self) -> boto3.client:
-
         """
         Authenticate MinIO credentials or AWS credentials and return an S3 client.
         This is a helper function called by read_path() or write_path().
         """
-        if self.minio_credentials:
-            s3_client = self.minio_credentials.get_s3_client()
-
-        elif self.aws_credentials:
-            s3_client = self.aws_credentials.get_s3_client()
-        else:
-            raise ValueError(
-                "S3 Bucket requires either a minio_credentials"
-                "field or an aws_credentials field."
-            )
-        return s3_client
+        return self.credentials.get_s3_client()
 
     def _get_bucket_resource(self) -> boto3.resource:
         """
         Retrieves boto3 resource object for the configured bucket
         """
-        if self.minio_credentials:
-            params_override = (
-                self.minio_credentials.aws_client_parameters.get_params_override()
-            )
-            if "endpoint_url" not in params_override and self.endpoint_url:
-                params_override["endpoint_url"] = self.endpoint_url
-            bucket = (
-                self.minio_credentials.get_boto3_session()
-                .resource("s3", **params_override)
-                .Bucket(self.bucket_name)
-            )
-
-        elif self.aws_credentials:
-            params_override = (
-                self.aws_credentials.aws_client_parameters.get_params_override()
-            )
-            bucket = (
-                self.aws_credentials.get_boto3_session()
-                .resource("s3", **params_override)
-                .Bucket(self.bucket_name)
-            )
-        else:
-            raise ValueError(
-                "S3 Bucket requires either a minio_credentials"
-                "field or an aws_credentials field."
-            )
+        params_override = self.credentials.aws_client_parameters.get_params_override()
+        bucket = (
+            self.credentials.get_boto3_session()
+            .resource("s3", **params_override)
+            .Bucket(self.bucket_name)
+        )
         return bucket
 
     @sync_compatible
@@ -478,7 +340,7 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
             local_path: Local path to download S3 contents to. Defaults to the current
                 working directory.
         """
-        bucket_folder = self.bucket_folder or self.basepath
+        bucket_folder = self.bucket_folder
         if from_path is None:
             from_path = str(bucket_folder) if bucket_folder else ""
 
@@ -556,7 +418,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
 
     @sync_compatible
     async def read_path(self, path: str) -> bytes:
-
         """
         Read specified path from S3 and return contents. Provide the entire
         path to the key in S3.
@@ -589,7 +450,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         return await run_sync_in_worker_thread(self._read_sync, path)
 
     def _read_sync(self, key: str) -> bytes:
-
         """
         Called by read_path(). Creates an S3 client and retrieves the
         contents from  a specified path.
@@ -598,7 +458,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         s3_client = self._get_s3_client()
 
         with io.BytesIO() as stream:
-
             s3_client.download_fileobj(Bucket=self.bucket_name, Key=key, Fileobj=stream)
             stream.seek(0)
             output = stream.read()
@@ -606,7 +465,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
 
     @sync_compatible
     async def write_path(self, path: str, content: bytes) -> str:
-
         """
         Writes to an S3 bucket.
 
@@ -645,7 +503,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         return path
 
     def _write_sync(self, key: str, data: bytes) -> None:
-
         """
         Called by write_path(). Creates an S3 client and uploads a file
         object.
@@ -654,7 +511,6 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         s3_client = self._get_s3_client()
 
         with io.BytesIO(data) as stream:
-
             s3_client.upload_fileobj(Fileobj=stream, Bucket=self.bucket_name, Key=key)
 
     # NEW BLOCK INTERFACE METHODS BELOW
@@ -690,7 +546,10 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
                 f"Bucket path {bucket_path!r} is already prefixed with "
                 f"bucket folder {self.bucket_folder!r}; is this intentional?"
             )
-        return (Path(self.bucket_folder) / bucket_path).as_posix()
+
+        return (Path(self.bucket_folder) / bucket_path).as_posix() + (
+            "" if not bucket_path.endswith("/") else "/"
+        )
 
     @sync_compatible
     async def list_objects(
@@ -855,7 +714,7 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         )
         self.logger.info(
             f"Downloaded object from bucket {self.bucket_name!r} path {bucket_path!r} "
-            f"to file object."
+            "to file object."
         )
         return to_file_object
 
@@ -927,6 +786,70 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
         await asyncio.gather(*async_coros)
 
         return Path(to_folder)
+
+    @sync_compatible
+    async def stream_from(
+        self,
+        bucket: "S3Bucket",
+        from_path: str,
+        to_path: Optional[str] = None,
+        **upload_kwargs: Dict[str, Any],
+    ) -> str:
+        """Streams an object from another bucket to this bucket.
+
+        Args:
+            bucket: The bucket to stream from.
+            from_path: The path of the object to stream.
+            to_path: The path to stream the object to. Defaults to the object's name.
+            **upload_kwargs: Additional keyword arguments to pass to
+                `Client.upload_fileobj`.
+
+        Returns:
+            The path that the object was uploaded to.
+
+        Examples:
+            Stream notes.txt from your-bucket/notes.txt to my-bucket/landed/notes.txt.
+
+            ```python
+            from prefect_aws.s3 import S3Bucket
+
+            your_s3_bucket = S3Bucket.load("your-bucket")
+            my_s3_bucket = S3Bucket.load("my-bucket")
+
+            my_s3_bucket.stream_from(
+                your_s3_bucket,
+                "notes.txt",
+                to_path="landed/notes.txt"
+            )
+            ```
+
+        """
+        if to_path is None:
+            to_path = Path(from_path).name
+
+        # Get the source object's StreamingBody
+        from_path: str = bucket._join_bucket_folder(from_path)
+        from_client = bucket.credentials.get_s3_client()
+        obj = await run_sync_in_worker_thread(
+            from_client.get_object, Bucket=bucket.bucket_name, Key=from_path
+        )
+        body: StreamingBody = obj["Body"]
+
+        # Upload the StreamingBody to this bucket
+        bucket_path = str(self._join_bucket_folder(to_path))
+        to_client = self.credentials.get_s3_client()
+        await run_sync_in_worker_thread(
+            to_client.upload_fileobj,
+            Fileobj=body,
+            Bucket=self.bucket_name,
+            Key=bucket_path,
+            **upload_kwargs,
+        )
+        self.logger.info(
+            f"Streamed s3://{bucket.bucket_name}/{from_path} to the bucket "
+            f"{self.bucket_name!r} path {bucket_path!r}."
+        )
+        return bucket_path
 
     @sync_compatible
     async def upload_from_path(
@@ -1025,7 +948,7 @@ class S3Bucket(WritableFileSystem, WritableDeploymentStorage, ObjectStorageBlock
             **upload_kwargs,
         )
         self.logger.info(
-            f"Uploaded from file object to the bucket "
+            "Uploaded from file object to the bucket "
             f"{self.bucket_name!r} path {bucket_path!r}."
         )
         return bucket_path