+++ prefect_aws/secrets_manager.py
@@ -332,9 +332,9 @@ async def delete_secret(
 
     delete_secret_kwargs: Dict[str, Union[str, int, bool]] = dict(SecretId=secret_name)
     if force_delete_without_recovery:
-        delete_secret_kwargs[
-            "ForceDeleteWithoutRecovery"
-        ] = force_delete_without_recovery
+        delete_secret_kwargs["ForceDeleteWithoutRecovery"] = (
+            force_delete_without_recovery
+        )
     else:
         delete_secret_kwargs["RecoveryWindowInDays"] = recovery_window_in_days
 
@@ -364,6 +364,10 @@ class AwsSecret(SecretBlock):
     """
     _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/1jbV4lceHOjGgunX15lUwT/db88e184d727f721575aeb054a37e277/aws.png?h=250"  # noqa
     _block_type_name = "AWS Secret"
+
+    _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/1jbV4lceHOjGgunX15lUwT/db88e184d727f721575aeb054a37e277/aws.png?h=250"  # noqa
+    _block_type_name = "AWS Secret"
+    _documentation_url = "https://prefecthq.github.io/prefect-aws/secrets_manager/#prefect_aws.secrets_manager.AwsSecret"  # noqa
 
     aws_credentials: AwsCredentials
     secret_name: str = Field(default=..., description="The name of the secret.")
@@ -488,7 +492,7 @@ class AwsSecret(SecretBlock):
             )
         elif not (7 <= recovery_window_in_days <= 30):
             raise ValueError(
-                f"Recovery window must be between 7 and 30 days, got "
+                "Recovery window must be between 7 and 30 days, got "
                 f"{recovery_window_in_days}."
             )